requirements.txt
asttokens==2.4.1
backcall==0.2.0
blinker==1.8.2
boto3==1.34.99
botocore==1.34.99
click==8.1.7
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
Flask==3.0.3
Flask-JWT-Extended==4.6.0
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jupyter_client==8.6.1
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.2.1
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.0.3
s3transfer==0.10.1
six==1.16.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.3
typing_extensions==4.11.0
urllib3==1.26.18
wcwidth==0.2.13
Werkzeug==3.0.3
zipp==3.18.1
